pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["Ivylou <princequincyy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
sqlalchemy = "^2.0.18"
psycopg2-binary = "^2.9.6"
alembic = "^1.11.1"
uvicorn = "^0.22.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
tenacity = "^8.2.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = {version = "1.*", extras = ["email"]}
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
httpx = "^0.24.1"
pytest = "^7.4.0"
emails = "^0.6"
jinja2 = "^3.1.2"
pytest-mock = "^3.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
